Backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
#  🌌 COSMIC AI - Backend Dependencies
# ═══════════════════════════════════════════════════════════════

# FastAPI Framework
fastapi
uvicorn[standard]
python-multipart

# Azure OpenAI
openai
sentence-transformers>=2.2.2

# Vector Database
faiss-cpu
numpy>=1.24.0

# BM25 Search & Reranking
rank-bm25==0.2.2
# spacy>=3.8.0
nltk>=3.8.1

# Document Processing
pypdf2
python-docx

# Text Chunking
langchain-text-splitters

# Utilities
python-dotenv
pydantic
pydantic-settings
requests

# Logging
colorama
rich
